+++ R/git.R
@@ -13,3 +13,8 @@ renv_git_preflight <- function() {
   if (!nzchar(Sys.which("git")))
     stopf("'git' is not available on the PATH")
 }
+
+renv_git_sha <- function(path) {
+  renv_scope_wd(path)
+  renv_system_exec("git", c("rev-parse", "HEAD"), action = "reading git commit")
+}